pyproject.toml
[tool.poetry]
name = "munzapp-example"
version = "0.1.0"
description = ""
authors = ["andrei_turobov <forgunn@yandex.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.1"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
django-ninja = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
